Move services into their own directories
I was creating test files for some services and it polluted the directory structure quite a lot. To clean up, I put each
service into its own directory.

I am doing this in a seperate commit so that git knows the files were moved, and doesn't do incorrectly say that the
entire file was changed.

[skip ci]
