+++ src/components/data-source/data-source.ts
@@ -4,10 +4,10 @@ import { customElement, property, query, state } from "lit/decorators.js";
 import { booleanConverter } from "../../helpers/attributes";
 import { downloadFile } from "../../helpers/files";
 import { AbstractComponent } from "../../mixins/abstractComponent";
-import { UrlSourcedFetcher } from "../../services/urlSourcedFetcher";
+import { UrlSourcedFetcher } from "../../services/urlSourcedFetcher/urlSourcedFetcher";
 import { LoadState, VerificationGridComponent } from "../verification-grid/verification-grid";
 import { required } from "../../helpers/decorators";
-import { PageFetcher } from "../../services/gridPageFetcher";
+import { PageFetcher } from "../../services/gridPageFetcher/gridPageFetcher";
 import { DownloadableResult } from "../../models/subject";
 import { when } from "lit/directives/when.js";
 import dataSourceStyles from "./css/style.css?inline";
@@ -116,18 +116,6 @@ export class DataSourceComponent extends AbstractComponent(LitElement) {
   private resultRows(): Partial<DownloadableResult>[] {
     if (!this.verificationGrid) {
       throw new Error("could not find verification grid component");
-    }
-
-    // If the data source is a URL, we know the entire dataset ahead of time.
-    // Because url data sources are static files that were probably provided by
-    // the user, we download the entire dataset with decisions added instead of
-    // just the subjects that the user has seen so far.
-    //
-    // This obeys our design principle of not changing the original data set and
-    // only adding information.
-    if (this.isUrlSourced()) {
-      const subjects = this.verificationGrid.readonlySubjects;
-      return subjects.map((model) => model.toDownloadable());
     }
 
     const subjects = this.verificationGrid.readonlySubjects;
