+++ src/helpers/files.ts
@@ -3,13 +3,20 @@
 // TODO: remove this once Firefox ESR supports the showSaveFilePicker API
 // https://caniuse.com/?search=showSaveFilePicker
 export async function downloadFile(file: File) {
-  const supportsSaveFile = "showSaveFilePicker" in window;
-  if (supportsSaveFile) {
-    const saveFileHandle = await (window as any).showSaveFilePicker({ suggestedName: file.name });
-    const writableStream = await saveFileHandle.createWritable();
+  if (typeof window.showSaveFilePicker === "function") {
+    try {
+      const saveFileHandle = await window.showSaveFilePicker({ suggestedName: file.name });
+      const writableStream = await saveFileHandle.createWritable();
 
-    await writableStream.write(file);
-    await writableStream.close();
+      await writableStream.write(file);
+      await writableStream.close();
+    } catch (error: any) {
+      // Fail silently if the user has simply canceled the dialog.
+      if (error.name !== "AbortError") {
+        console.error(error.name, error.message);
+        return;
+      }
+    }
   } else {
     const url = URL.createObjectURL(file);
     const a = document.createElement("a");
